+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/scene/resource_widget/private/media_resource_widget_p.cpp
@@ -372,7 +372,7 @@ const char* const MediaResourceWidgetPrivate::motionSkipMask(int channel) const
     static const MotionSkipMask kEmptyMask = QnRegion();
 
     ensureMotionSkip();
-    return channel < m_motionSkipMaskCache->size()
+    return channel < (int) m_motionSkipMaskCache->size()
         ? m_motionSkipMaskCache->at(channel).bitMask()
         : kEmptyMask.bitMask();
 }
