+++ vms/client/nx_vms_client_desktop/src/nx/vms/client/desktop/system_administration/models/user_list_model.cpp
@@ -340,7 +340,7 @@ Qt::CheckState UserListModel::Private::checkState() const
         return Qt::Unchecked;
 
     if ((checkedUsers.size() + systemContext()->nonEditableUsersAndGroups()->userCount())
-        == users.size())
+        == (int) users.size())
     {
         return Qt::Checked;
     }
